Refactor: 주문 처리 로직 개선
