+++ auto_stock/trading/tasks/__init__.py
@@ -1,4 +1,3 @@
-from .auto_buy import *
-from .auto_sell import *
+from .auto_order import *
 from .debug import *
-from .trade_process import *
+from .auto_re_order import *
